requirements.txt
Flask
Flask-WTF
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
WeasyPrint
email-validator
reportlab
mysqlclient
